ml_pipeline/requirements.txt
fastf1
pandas
numpy
xgboost
scikit-learn
supabase
python-dotenv
fastapi>=0.110
uvicorn[standard]>=0.27
